refactor test_gof in a separate file
